Add files via upload
Whale_identification_conference_template_letter.docx is the file we can write our info in. The initial things I added so far are in navy blue just so it's easier to tell what has been changed from the original template. Right now text from the original template is in black and is kept in place to keep the formatting consistent with the original template. The original template file is conference-template-letter.docx.